refactor: introduce BookData structure and update BooksModule to use new configuration
